Fix otherPrint and TestEmptyInput tests
There were a lot of abstract classes defined in Rendering/Core or
Rendering/Volume that the OpenGL2 backend has not implemented yet.
This causes a crach whenever such a class is instantiated as is
done by some of the tests. The fix is to either remove or provide
implementations for these classes. For the volme code this patch adds
a VolumeOpenGL2 directory which will be where the new/converted
volume rendering implementations will go.

The other big piece is the painters. As we are not/can not support that API
the classes were moved to OpenGL. The excpetion is the PainterDeviceAdapter
which is part of RenderWindow. When building with OpenGL2 that class is
treated in RenderWindow as a plain vtkObject that does nothing.

vtkSmartVolumeMapper is not currenty built by OpenGL2 but could be if we
are willing to add a bunch of ifdefs. The basic approach is there already
in the class, just need to add a bunch more if we want to have that in the
short term prior to the new GPU mapper being completed.

This patch includes some other fixes that were required to get the tests to
pass. Generally related to missing file properties or extra includes that
were not needed.

+++ Rendering/Volume/vtkSmartVolumeMapper.cxx
@@ -20,7 +20,6 @@
 #include "vtkDataArray.h"
 #include "vtkFixedPointVolumeRayCastMapper.h"
 #include "vtkEventForwarderCommand.h"
-#include "vtkGPUVolumeRayCastMapper.h"
 #include "vtkImageData.h"
 #include "vtkImageResample.h"
 #include "vtkPiecewiseFunction.h"
@@ -30,6 +29,8 @@
 #include "vtkVolumeProperty.h"
 #include "vtkVolumeTextureMapper3D.h"
 #include <cassert>
+
+#include "vtkGPUVolumeRayCastMapper.h"
 
 vtkStandardNewMacro( vtkSmartVolumeMapper );
 
@@ -55,12 +56,20 @@ vtkSmartVolumeMapper::vtkSmartVolumeMapper()
 
   // Create all the mappers we might need
   this->RayCastMapper   = vtkFixedPointVolumeRayCastMapper::New();
+
+// opengl2 back end only has fixed point mapper right now
+#ifdef  VTK_OPENGL2
+  this->GPUMapper       = NULL;
+  this->TextureMapper   = NULL;
+  this->GPULowResMapper = NULL;
+#else
   this->GPUMapper       = vtkGPUVolumeRayCastMapper::New();
   this->MaxMemoryInBytes=this->GPUMapper->GetMaxMemoryInBytes();
   this->MaxMemoryFraction=this->GPUMapper->GetMaxMemoryFraction();
 
   this->TextureMapper   = vtkVolumeTextureMapper3D::New();
   this->GPULowResMapper = vtkGPUVolumeRayCastMapper::New();
+#endif
 
   // If the render window has a desired update rate of at least 1 frame
   // per second or more, we'll consider this interactive
@@ -97,6 +106,7 @@ vtkSmartVolumeMapper::vtkSmartVolumeMapper()
   this->RayCastMapper->AddObserver(vtkCommand::VolumeMapperComputeGradientsProgressEvent, cb);
 
   // And the texture mapper's events
+#ifndef  VTK_OPENGL2
   this->TextureMapper->AddObserver(vtkCommand::StartEvent, cb);
   this->TextureMapper->AddObserver(vtkCommand::EndEvent, cb);
   this->TextureMapper->AddObserver(vtkCommand::ProgressEvent, cb);
@@ -106,6 +116,7 @@ vtkSmartVolumeMapper::vtkSmartVolumeMapper()
   this->TextureMapper->AddObserver(vtkCommand::VolumeMapperComputeGradientsStartEvent, cb);
   this->TextureMapper->AddObserver(vtkCommand::VolumeMapperComputeGradientsEndEvent, cb);
   this->TextureMapper->AddObserver(vtkCommand::VolumeMapperComputeGradientsProgressEvent, cb);
+#endif
 
   // And the GPU mapper's events
   // Commented out because too many events are being forwwarded
@@ -133,11 +144,31 @@ vtkSmartVolumeMapper::vtkSmartVolumeMapper()
 // ----------------------------------------------------------------------------
 vtkSmartVolumeMapper::~vtkSmartVolumeMapper()
 {
-  this->RayCastMapper->Delete();
-  this->GPUMapper->Delete();
-  this->GPULowResMapper->Delete();
-  this->TextureMapper->Delete();
-  this->GPUResampleFilter->Delete();
+  if (this->RayCastMapper)
+    {
+    this->RayCastMapper->Delete();
+    this->RayCastMapper = 0;
+    }
+  if (this->GPUMapper)
+    {
+    this->GPUMapper->Delete();
+    this->GPUMapper = 0;
+    }
+  if (this->GPULowResMapper)
+    {
+    this->GPULowResMapper->Delete();
+    this->GPULowResMapper = 0;
+    }
+  if (this->TextureMapper)
+    {
+    this->TextureMapper->Delete();
+    this->TextureMapper = 0;
+    }
+  if (this->GPUResampleFilter)
+    {
+    this->GPUResampleFilter->Delete();
+    this->GPUResampleFilter = 0;
+    }
 }
